Project Start
See MongoLab.txt for database info. In addition to the basic stack, I
added an applications crud module (very similar to articles). Most of
the work should be done in the application directories. I also edited
the home page and started the application layout.  Contact me if you
have questions. Lets do this.